+++ Pick/src/test/java/com/nob/pick/post/DeletePostTest.java
@@ -77,7 +77,7 @@ public class DeletePostTest {
 			targetPost != null ?
 			targetPost.getStatus().toString() : "NULL"
 		);
-		int memberId = (
+		Long memberId = (
 			targetPost != null ?
 			targetPost.getMember().getMemberId() : 0
 		);
